feat(distri-fiche 13A.1+A.2): extraire 8 blocs partages + refactor modale
Chantier #13 sous-vague A — Fiche distri unifiee. Etapes 13A.1 + 13A.2.

Objectif : extraire les 8 blocs de TeamMemberDrilldownModal en composants
partages dans src/components/distributor-blocks/, qui seront reutilises
par la nouvelle page enrichie /distributors/:id (etape 13A.3 suivante).

NOUVEAU — src/components/distributor-blocks/
  - _shared.tsx       : SectionTitle, MetricCard, BreakdownRow, PvTierRow,
                        PrimaryActionButton, AdminCard, formatRelativeFR,
                        styles partages (grids, hint, label)
  - EngagementTotalBlock : XP total + level + breakdown 6 categories
  - ApprentissageBlock   : Academy + Formation cards
  - ActiviteRecenteBlock : Bilans 30j + RDV 30j + Messages 7j
  - EngagementBlock      : Derniere connexion + Connexions cumulees
  - RangHerbalifeBlock   : admin select rang + Appliquer (state autonome)
  - ProgressionRangBlock : jauge progression vers prochain rang (auto-calc
                           qualifyingPersonalPv via context.users +
                           usePvBreakdowns)
  - PvBizworksBlock      : 5 PvTierRow + total live + override estim. +
                           Appliquer (state autonome, hydrate
                           usePvBreakdowns)
  - CompteActifBlock     : toggle Geler/Reactiver
  - index.ts             : barrel exports

Chaque bloc admin est AUTONOME : embarque son state, son toast, son RPC
call. Pour partager le refresh entre la modale et la page, ils exposent
une prop onApplied? (callback declenchee apres succes).

REFACTOR — TeamMemberDrilldownModal
  - 1092 lignes -> 264 lignes (-76%)
  - Importe les 8 blocs et les compose conditionnellement selon les
    permissions (isAdmin && !isSelf)
  - Conserve : header avatar/nom/role/status + footer CTAs + overlay/modal
    styles
  - Aucune regression fonctionnelle attendue : les blocs reproduisent
    exactement le JSX/styles/handlers d'origine, juste deplaces.

BUILD : npm run build OK.

PROCHAINE ETAPE 13A.3 : construire la page enrichie /distributors/:id
avec 5 onglets (Vue d'ensemble / Clients / PV & Rentabilite / Activite /
Parametres distri) qui consomme ces memes blocs.
